Split check and explain integrated tests
