+++ fortitude/tests/explain.rs
@@ -3,18 +3,6 @@ use predicates::prelude::*;
 use std::process::Command;
 
 #[test]
-fn check_file_doesnt_exist() -> anyhow::Result<()> {
-    let mut cmd = Command::cargo_bin("fortitude")?;
-
-    cmd.arg("check").arg("test/file/doesnt/exist");
-    cmd.assert()
-        .failure()
-        .stdout(predicate::str::contains("No such file"));
-
-    Ok(())
-}
-
-#[test]
 fn explain_all() -> anyhow::Result<()> {
     let mut cmd = Command::cargo_bin("fortitude")?;
 
@@ -108,3 +96,4 @@ fn explain_mixed_multiple() -> anyhow::Result<()> {
 
     Ok(())
 }
+
